requirements.txt
# 🐢 Core Dependencies
groq>=0.4.0
pydantic>=2.0.0
python-dateutil>=2.8.2
rich>=13.0.0

# 🔐 Google API Integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0

# 🧠 Optimization
ortools>=9.0.0

# 🧪 Testing
pytest>=7.4.0
pytest-mock>=3.12.0
